chore: split actions-events test into separate test files
